+++ src/main/java/org/c4marathon/assignment/domain/type/TransactionState.java
@@ -1,0 +1,7 @@
+package org.c4marathon.assignment.domain.type;
+
+public enum TransactionState {
+	PENDING,
+	FINISHED,
+	CANCELLED
+}
